refactor(store): extract schema DDL into store/schema.py
