+++ tests/test_e2e_integration.py
@@ -40,12 +40,12 @@ class TestDockerfileCron:
 
 class TestDatabaseSchema:
     def test_database_has_lea_reports_table(self):
-        with open(os.path.join(PROJECT_ROOT, "database.py")) as f:
+        with open(os.path.join(PROJECT_ROOT, "store", "schema.py")) as f:
             content = f.read()
         assert "CREATE TABLE IF NOT EXISTS lea_reports" in content
 
     def test_database_has_ops_snapshots_table(self):
-        with open(os.path.join(PROJECT_ROOT, "database.py")) as f:
+        with open(os.path.join(PROJECT_ROOT, "store", "schema.py")) as f:
             content = f.read()
         assert "CREATE TABLE IF NOT EXISTS ops_snapshots" in content
